+++ deprecated.go
@@ -19,6 +19,9 @@ var IsAdmin = IsRoot
 // Deprecated: Use Rename() instead. This function may be removed in the future.
 var ReplaceFile = Rename
 
+// Deprecated: Use Options instead. This may be removed in the future.
+type FileOptions = Options
+
 // Deprecated: Use Supports*() functions instead. This may be removed in the future.
 type SupportedType = supportsType