+++ writefile_atomic_test.go
@@ -107,7 +107,7 @@ func TestWriteFileAtomicMode(t *testing.T) {
 	perm644 := os.FileMode(0o644)
 	perm600 := os.FileMode(0o600)
 
-	err := compat.WriteFileAtomic(file, content, compat.FileMode(perm644))
+	err := compat.WriteFileAtomic(file, content, compat.UseFileMode(perm644))
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}
@@ -132,7 +132,7 @@ func TestWriteFileAtomicMode(t *testing.T) {
 		t.Fatalf("Failed to change file mode: %q: %v", file, err)
 	}
 
-	err = compat.WriteFileAtomic(file, content, compat.FileMode(perm644))
+	err = compat.WriteFileAtomic(file, content, compat.UseFileMode(perm644))
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}