+++ writereader_atomic_test.go
@@ -108,7 +108,7 @@ func TestWriteReaderAtomicMode(t *testing.T) {
 	perm644 := os.FileMode(0o644)
 	perm600 := os.FileMode(0o600)
 
-	err := compat.WriteReaderAtomic(file, content, compat.FileMode(perm644))
+	err := compat.WriteReaderAtomic(file, content, compat.UseFileMode(perm644))
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}
@@ -133,7 +133,7 @@ func TestWriteReaderAtomicMode(t *testing.T) {
 		t.Fatalf("Failed to change file mode: %q: %v", file, err)
 	}
 
-	err = compat.WriteReaderAtomic(file, content, compat.FileMode(perm644))
+	err = compat.WriteReaderAtomic(file, content, compat.UseFileMode(perm644))
 	if err != nil {
 		t.Fatalf("Failed to write file: %q: %v", file, err)
 	}